(WIP) config: coreboot-x230: add gbe ifd and me blobs and increase cbfs size
WARNING: Updating from inside of Heads is NOT possible! Changing the IFD
format requires flashing from the outside (once).

This is because we only flash the bios region and whatever IFD format is
_currently_ in flash is used again. For this update, more regions need to
get flashed. Either:
 * flash ifd, me and bios regions from userspace manually, or
 * split up a new coreboot build to 4M and 8M and flash externally

This requires blobs for the x230 similar to the x220 now. They must be
modified by me_cleaner to be able to have 0xbe5000 of CBFS size. An
extract.sh script is provided, to be run before building Heads.

This should create enough room to keep the x230 supported by Heads in the
long run (when vboot or more tools might be desired).

add extract.sh